+++ NEWS.md
@@ -1,3 +1,18 @@
+# ggpubr 0.2.1.999
+  
+## New fatures
+
+
+## Major changes
+  
+  
+## Minor changes
+   
+
+## Bug fixes
+   
+
+
 # ggpubr 0.2.1
    
 ## Minor changes